@@ -14,7 +14,9 @@ namespace Compos.Coreforce
             string password,
             string clientId,
             string clientSecret,
-            string apiVersion = "v43.0"
+            string apiVersion = "v43.0",
+            bool enableCache = false,
+            int cacheExpirationInSeconds = 30
             )
         {
             // Set credentials
@@ -24,6 +26,8 @@ namespace Compos.Coreforce
             CoreforceConfiguration.Password = password;
             CoreforceConfiguration.Username = username;
             CoreforceConfiguration.ApiVersion = apiVersion;
+            CoreforceConfiguration.IsCacheEnabled = enableCache;
+            CoreforceConfiguration.CacheExpirationInSeconds = cacheExpirationInSeconds;
 
             if (environment == Environment.Production)
                 CoreforceConfiguration.AuthorizationUrl = "https://login.salesforce.com/services/oauth2/token";
