@@ -1,5 +1,7 @@
-﻿using Compos.Coreforce.Extensions;
+﻿using Compos.Coreforce.Attributes;
+using Compos.Coreforce.Extensions;
 using System;
+using System.Globalization;
 using System.Linq.Expressions;
 
 namespace Compos.Coreforce.Filter
@@ -10,6 +12,26 @@ namespace Compos.Coreforce.Filter
         private readonly object value;
         private readonly string filterOperator;
         private readonly FilterConcatination? concatination;
+
+        public Expression<Func<T, object>> Field
+        {
+            get => field;
+        }
+
+        public object Value
+        {
+            get => value;
+        }
+
+        public string FilterOperator
+        {
+            get => filterOperator;
+        }
+
+        public FilterConcatination? Concatination
+        {
+            get => concatination;
+        }
 
         public Filter(
             Expression<Func<T, object>> field, 
@@ -51,14 +73,25 @@ namespace Compos.Coreforce.Filter
             else if (expression.Type == typeof(string))
                 compareString = $"'{value}'";
 
+            else if (expression.Type == typeof(Date?) || expression.Type == typeof(Date))
+            {
+                var date = new Date(value.ToString());
+                var dateAsString = date.GetDate();
+
+                compareString = $"{dateAsString}";
+            }
+
             else if (expression.Type == typeof(DateTime?) || expression.Type == typeof(DateTime))
             {
                 var dateTime = Convert.ToDateTime(value).ToUniversalTime();
                 compareString = $"{dateTime.Year}-{dateTime.Month.ToString("D2")}-{dateTime.Day.ToString("D2")}T{dateTime.Hour.ToString("D2")}:{dateTime.Minute.ToString("D2")}:{dateTime.Second.ToString("D2")}Z";
             }
 
+            else if (value.GetType() == typeof(bool?) || value.GetType() == typeof(bool))
+                compareString = $"{value}";
+
             else
-                compareString = $"{value.ToString().ToLower()}";
+                compareString = $"{(value as IFormattable).ToString(null, CultureInfo.InvariantCulture).ToLower()}";
 
             if (concatination.HasValue && concatination == FilterConcatination.And)
                 return $"+{expression.Member.Name}+{filterOperator}+{compareString}+AND";
@@ -75,6 +108,26 @@ namespace Compos.Coreforce.Filter
         private readonly object value;
         private readonly string filterOperator;
         private readonly FilterConcatination? concatination;
+
+        public string Field
+        {
+            get => field;
+        }
+
+        public object Value
+        {
+            get => value;
+        }
+
+        public string FilterOperator
+        {
+            get => filterOperator;
+        }
+
+        public FilterConcatination? Concatination
+        {
+            get => concatination;
+        }
 
         public Filter(
             string field,
@@ -111,14 +164,25 @@ namespace Compos.Coreforce.Filter
             else if (value.GetType() == typeof(string))
                 compareString = $"'{value}'";
 
+            else if(value.GetType() == typeof(Date?) || value.GetType() == typeof(Date))
+            {
+                var date = new Date(value.ToString());
+                var dateAsString = date.GetDate();
+
+                compareString = $"{dateAsString}";
+            }
+
             else if (value.GetType() == typeof(DateTime?) || value.GetType() == typeof(DateTime))
             {
                 var dateTime = Convert.ToDateTime(value).ToUniversalTime();
                 compareString = $"{dateTime.Year}-{dateTime.Month.ToString("D2")}-{dateTime.Day.ToString("D2")}T{dateTime.Hour.ToString("D2")}:{dateTime.Minute.ToString("D2")}:{dateTime.Second.ToString("D2")}Z";
             }
 
+            else if(value.GetType() == typeof(bool?) || value.GetType() == typeof(bool))
+                compareString = $"{value}";
+
             else
-                compareString = $"{value.ToString().ToLower()}";
+                compareString = $"{(value as IFormattable).ToString(null, CultureInfo.InvariantCulture).ToLower()}";
 
             if (concatination.HasValue && concatination == FilterConcatination.And)
                 return $"+{field}+{filterOperator}+{compareString}+AND";
